% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiopacity.R
\name{addDynamicOpacityControls}
\alias{addDynamicOpacityControls}
\title{Add Dynamic Opacity Controls}
\usage{
addDynamicOpacityControls(
  map,
  type = c("category", "group", "layerId"),
  category = NULL,
  group = NULL,
  layerId = NULL,
  collapsed = FALSE,
  position = c("topright", "topleft", "bottomright", "bottomleft"),
  title = NULL
)
}
\arguments{
\item{map}{The map to add the opacity controls to.}

\item{category}{One or more categories to render opacity controls to.}

\item{group}{One or more groups to render opacity controls to.}

\item{layerId}{One or more layer IDs to render opacity controls to.}

\item{collapsed}{If FALSE (the default), the opacity control will always appear
in its expanded state. Set to TRUE to have the opacity control
rendered as an icon that expands when hovered over.}

\item{position}{Position of control: "topleft", "topright", "bottomleft", or "bottomright".}

\item{title}{The control title.}
}
\value{

}
\description{
Add Dynamic Opacity Controls
}
