% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multiopacity.R
\name{addOpacityControls}
\alias{addOpacityControls}
\title{Add Opacity Controls}
\usage{
addOpacityControls(
  map,
  layerId = NULL,
  collapsed = FALSE,
  position = c("topright", "topleft", "bottomright", "bottomleft"),
  label = NULL
)
}
\arguments{
\item{map}{The map to add the opacity controls to.}

\item{layerId}{One or more layer IDs to render opacity controls to. If NULL (the default),
will create controls for all layers and graphical components.}

\item{collapsed}{If FALSE (the default), the opacity control will always appear
in its expanded state. Set to TRUE to have the opacity control
rendered as an icon that expands when hovered over.}

\item{position}{Position of control: "topleft", "topright", "bottomleft", or "bottomright".}

\item{label}{The control title.}
}
\description{
Add Opacity Controls
}
\examples{

}
