% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{dependencies}
\alias{dependencies}
\title{Return JavaScript dependencies}
\usage{
dependencies()
}
\value{
List with JavaScript dependencies of the package.
}
\description{
Return JavaScript dependencies
}
