% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{leafletDebug}
\alias{leafletDebug}
\title{Leaflet Debug}
\usage{
leafletDebug(map)
}
\arguments{
\item{map}{Leaflet map.}
}
\value{
Leaflet map.
}
\description{
Debug a leaflet map in client side using browser inspector tools.
}
