% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{registerPlugin}
\alias{registerPlugin}
\title{Register Plugin to leaflet map}
\usage{
registerPlugin(map, plugin)
}
\arguments{
\item{map}{The map to register the plugin.}

\item{plugin}{List with JavaScript dependencies (`from htmltools::htmlDependency`)
to register.}
}
\value{
Leaflet map with plugin registered (JS dependencies added).
}
\description{
Register Plugin to leaflet map
}
